+++ src/components/Projects/FilteringMenu.js
@@ -19,11 +19,7 @@ const FilteringMenu = ({ projectData, searchText, setSearchText, selectedChips, 
         Filter by technology
       </Typography>
 
-      <FilteringBox
-        uniqueTools={uniqueTools}
-        selectedChips={selectedChips}
-        setSelectedChips={setSelectedChips}
-      />
+      <FilteringBox uniqueTools={uniqueTools} selectedChips={selectedChips} setSelectedChips={setSelectedChips} />
 
       <Divider sx={{ margin: '2rem' }} />
     </>
