+++ src/components/Projects/ProjectsContainer.js
@@ -1,28 +1,11 @@
-import { useState, useEffect } from 'react';
+import React from 'react';
 import FilteringMenu from './FilteringMenu';
 import ProjectCards from './ProjectCards';
-import { useFilter } from './useFilter';
+import useProjectFilter from './useProjectFilter';
 
 const ProjectsContainer = ({ projectData }) => {
-  const [filteredProjects, setFilteredProjects] = useState([]);
-  const [searchText, setSearchText] = useState('');
-  const [selectedChips, setSelectedChips] = useState(new Set());
-
-  const { filterByUniqueProjects, filterBySelectedChips } = useFilter();
-
-  // Extracts the unique tools used across all projects
-  const uniqueTools = filterByUniqueProjects({ projectData });
-
-  // Filter projects based on selected chips
-  const chipFilteredProjects = filterBySelectedChips({ selectedChips, projectData });
-
-  // Update filtered projects whenever search text or selected chips change
-  useEffect(() => {
-    const searchedProjects = chipFilteredProjects.filter((project) =>
-      project.title.toLowerCase().includes(searchText.toLowerCase())
-    );
-    setFilteredProjects(searchedProjects);
-  }, [searchText, selectedChips, chipFilteredProjects]);
+  const { filteredProjects, searchText, setSearchText, selectedChips, setSelectedChips, uniqueTools } =
+    useProjectFilter(projectData);
 
   return (
     <>
